adicionando seções de informações da compra
